Refactor to initialize the DelegationIdentity only once.
